Use the user-configured nixpkgs
We pinned nixpkgs before, but this leads to problems when Python modules
from this package set are mixed with modules from nixpkgs, leading to
incompatible versions of libstdc++, etc.

nixpkgs now uses GCC 9, so we override Tensile with a small patch (from
nixos-rocm) to fix the build. Also properly compose the overlays to
ensure that our own overrides are picked up by other derivations from
nixos-rocm.